completed all queries

+++ app/db/models/city.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Numeric, ForeignKey
+from sqlalchemy import Column, Integer, String, Numeric, ForeignKey, Float
 from sqlalchemy.orm import relationship
 
 from app.db.models import Base
@@ -10,10 +10,10 @@ class City(Base):
     city_id = Column(Integer, primary_key=True, autoincrement=True)
     city_name = Column(String(100), nullable=True)
     country_id = Column(Integer, ForeignKey('countries.country_id'), nullable=False)
-    latitude = Column(Numeric, nullable=True)
-    longitude = Column(Numeric, nullable=True)
+    latitude = Column(Float, nullable=True)
+    longitude = Column(Float, nullable=True)
 
-    # Relationship to the Country model (if you have one)
+
     country = relationship("Country", back_populates="cities")
 
     def __repr__(self):
